eventstream: split client.go, refactor queryObjectsApi*
As the eventstream.Client code grew, the client.go file became too long
for my liking. Thus, I split it into main/common, Event Stream, and
other Icinga 2 API parts.

Furthermore, preparing future changes, I refactored the queryObjectsApi*
functions to allow direct object access without a "slow" filter.
